fix: handle provider state better
Previously, the fetch error during the init phase didn't allowed to setup
polling. It could be that next fetch will be successful and the fresh
values will be returned.

The polling is reworked a bit. Now the ready state provider will become stale
is there is any error during the fetch. Also the stale or error state provider
will recover to ready.